recompiled fastpd to work with longer ints - seems to be working

+++ FastPD_DemoVersion/src/graph.h
@@ -115,7 +115,7 @@ public:
 
 	/* Type of edge weights.
 	   Can be changed to char, int, float, double, ... */
-	typedef int Real;
+	typedef long Real;
 	typedef Real captype;
 
 	/* Type of total flow */
